refactor component project structure
