+++ frontend/src/components/Home/MoveSearchBar/styles.ts
