+++ frontend/src/components/Home/MovesList/MoveCard/MoveCardBody/styles.ts
