+++ frontend/src/components/Home/MovesList/MoveCard/MoveCardHeader/styles.ts
