+++ frontend/src/components/Home/MovesList/MoveCard/styles.ts
