+++ frontend/src/components/Home/MovesList/styles.ts
