+++ frontend/src/components/PokemonResults/PokemonCardList/PokemonCard/PokemonCardBody/styles.ts
