+++ frontend/src/components/PokemonResults/PokemonCardList/PokemonCard/PokemonCardHeader/styles.ts
