+++ frontend/src/components/PokemonResults/PokemonCardList/PokemonCard/styles.ts
